Move GoldMiner::MessagesQuery to GoldMiner::Slack::MessagesQuery
