+++ spec/gold_miner/messages_query_spec.rb
@@ -2,10 +2,10 @@
 
 require "spec_helper"
 
-RSpec.describe GoldMiner::MessagesQuery do
+RSpec.describe GoldMiner::Slack::MessagesQuery do
   describe "#on_channel" do
     it "sets the channel to search messages in" do
-      query = GoldMiner::MessagesQuery.new
+      query = described_class.new
 
       result = query.on_channel("dev")
 
@@ -15,7 +15,7 @@ RSpec.describe GoldMiner::MessagesQuery do
 
   describe "#sent_after" do
     it "sets the start date to search messages" do
-      query = GoldMiner::MessagesQuery.new
+      query = described_class.new
 
       result = query.sent_after("2018-01-01")
 
@@ -25,7 +25,7 @@ RSpec.describe GoldMiner::MessagesQuery do
 
   describe "#with_reaction" do
     it "sets the query reaction to search messages" do
-      query = GoldMiner::MessagesQuery.new
+      query = described_class.new
 
       result = query.with_reaction("thumbsup")
 
@@ -35,7 +35,7 @@ RSpec.describe GoldMiner::MessagesQuery do
 
   describe "#with_topic" do
     it "sets the query topic to TIL messages" do
-      query = GoldMiner::MessagesQuery.new
+      query = described_class.new
 
       result = query.with_topic("TIL")
 
@@ -45,7 +45,7 @@ RSpec.describe GoldMiner::MessagesQuery do
 
   describe "#to_s" do
     it "returns the string representation of the query" do
-      query = GoldMiner::MessagesQuery.new
+      query = described_class.new
 
       result = query
         .with_topic("TIL")
@@ -58,7 +58,7 @@ RSpec.describe GoldMiner::MessagesQuery do
     end
 
     it "does not include unset options" do
-      query = GoldMiner::MessagesQuery.new
+      query = described_class.new
 
       result = query.to_s
